requirements.txt
streamlit>=1.30.0
openai>=1.30.0
pydantic>=2.0.0
pypdf>=4.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
plotly>=5.18.0
streamlit[auth]>=1.30.0
authlib>=1.3.2
psycopg[binary]>=3.1.0
